Move test specific utils into their own file
