+++ src/utils.ts
@@ -1,72 +1,8 @@
 import * as crypto from 'crypto';
 import * as fs from 'fs';
-import { normalize } from 'jest-config';
-import * as setFromArgv from 'jest-config/build/set_from_argv';
 import * as path from 'path';
 import * as tsc from 'typescript';
 import { JestConfig, TsJestConfig } from './jest-types';
-
-function parseConfig(argv) {
-  if (argv.config && typeof argv.config === 'string') {
-    // If the passed in value looks like JSON, treat it as an object.
-    if (argv.config[0] === '{' && argv.config[argv.config.length - 1] === '}') {
-      return JSON.parse(argv.config);
-    }
-  }
-  return argv.config;
-}
-
-function loadJestConfigFromFile(filePath, argv) {
-  const config = JSON.parse(fs.readFileSync(filePath, 'utf-8'));
-  config.rootDir = config.rootDir
-    ? path.resolve(path.dirname(filePath), config.rootDir)
-    : process.cwd();
-  return normalize(config, argv);
-}
-
-function loadJestConfigFromPackage(filePath, argv) {
-  /* tslint:disable */
-  const R_OK = (fs.constants && fs.constants.R_OK) || (fs['R_OK'] as number);
-  /* tslint:enable */
-  try {
-    fs.accessSync(filePath, R_OK);
-  } catch (e) {
-    return null;
-  }
-
-  const packageData = require(filePath);
-  const config = packageData.jest || {};
-  const root = path.dirname(filePath);
-  config.rootDir = config.rootDir ? path.resolve(root, config.rootDir) : root;
-  return normalize(config, argv);
-}
-
-function readRawConfig(argv, root) {
-  const rawConfig = parseConfig(argv);
-
-  if (typeof rawConfig === 'string') {
-    return loadJestConfigFromFile(path.resolve(process.cwd(), rawConfig), argv);
-  }
-
-  if (typeof rawConfig === 'object') {
-    const config = Object.assign({}, rawConfig);
-    config.rootDir = config.rootDir || root;
-    return normalize(config, argv);
-  }
-
-  const packageConfig = loadJestConfigFromPackage(
-    path.join(root, 'package.json'),
-    argv,
-  );
-  return packageConfig || normalize({ rootDir: root }, argv);
-}
-
-export function getJestConfig(root) {
-  const yargs = require('yargs');
-  const argv = yargs(process.argv.slice(2)).argv;
-  const rawConfig = readRawConfig(argv, root);
-  return Object.freeze(setFromArgv.default(rawConfig, argv));
-}
 
 export function getTSJestConfig(globals: any): TsJestConfig {
   return globals && globals['ts-jest'] ? globals['ts-jest'] : {};
